ceph: more mgr autoscaler option
Now that the pg_autoscaler is on by default, it is "normal" (and okay) to
have a small number of PGs in the cluster if the overall cluster usage is
also low. This setting just results in a health warning out of the box
when you create a pool and haven't written any data yet.

fyi: ceph upstream https://github.com/ceph/ceph/pull/30352

Signed-off-by: Sébastien Han <seb@redhat.com>